fastApiProject/pyproject.toml
[build-system]
requires = ["setuptools>=75.3.0"]
build-backend = "setuptools.build_meta"

[project]
name = "Email_Sender_Service"
version = "0.1.0"
description = "From Kafka recives email data and then gets their data using django rest framework and sends email using smtp"
authors = [{name = "Nima Hekmati", email = "hekmatinima@gmail.com"}]
requires-python = "~=3.12"
dependencies = [
    "fastapi~=0.115.8",
    "pydantic==2.4.0",
    "python-dotenv~=1.0.1",
    "uvicorn==0.34.0",
    "python-decouple~=3.8",
    "aiokafka~=0.12.0",
    "kafka-python~=2.0.4",
    "typing_extensions~=4.12.2",
    "starlette~=0.45.3",
    "djangorestframework~=3.15.2",
    "requests~=2.32.3"
]

[project.optional-dependencies]
dev = [
    "pytest~=8.3.4",
    "pytest-asyncio==0.25.3",
    "uv==0.6.3"
]

[tool.uv]
index-url = "https://pypi.org/simple/"
